Extract listing-PR workflow into its own handler
Address review feedback that command-runner.ts is meant to be
command-agnostic but had hardcoded knowledge of pr-listing-create
and pr-listing-retry (constants, types, step methods, GitHub/lint
dependencies).

- Add BotCommandHandler interface and makeEnqueueRunCommand factory
  to command-runner.ts; CommandRunner now takes handlers[] and
  dispatches to the first match (or falls back to generic enqueue).
- Move all listing-PR specifics — command strings, event types,
  WorkflowContext, StepError, step methods, helpers — into a new
  pr-listing-workflow-handler.ts implementing BotCommandHandler.
- Wire PrListingWorkflowHandler into timeline-handler.ts.
- Test setup goes through a small makeRunner() helper.
